pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "Django todo-list app"
authors = ["TheArtur128 <88573504+TheArtur128@users.noreply.github.com>"]
license = "Apache-2.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.10"
act4 = "^3.2.0"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
redis = "^5.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
event-bus = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-django = "^4.7.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
ignore-init-module-imports = true
exclude = ["*/migrations/*"]

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["apps/*/tests.py", "apps/*/tests/test_*.py"]
